[MPLUGIN-442] Generate goal documentation leveraging Sink API
Drop XDoc intermediate format
