+++ maven-plugin-report-plugin/pom.xml
@@ -77,10 +77,6 @@
       <groupId>org.apache.maven.plugin-tools</groupId>
       <artifactId>maven-plugin-tools-api</artifactId>
     </dependency>
-    <dependency>
-      <groupId>org.apache.maven.plugin-tools</groupId>
-      <artifactId>maven-plugin-tools-generators</artifactId>
-    </dependency>
 
     <!-- doxia -->
     <dependency>
@@ -95,6 +91,12 @@
       <artifactId>maven-reporting-impl</artifactId>
     </dependency>
 
+    <dependency>
+      <groupId>org.apache.commons</groupId>
+      <artifactId>commons-lang3</artifactId>
+      <version>3.8.1</version>
+      <!--same version as used by doxia site renderer -->
+    </dependency>
     <!-- plexus -->
     <dependency>
       <groupId>org.codehaus.plexus</groupId>
