+++ maven-plugin-report-plugin/src/main/resources/plugin-report.properties
@@ -28,6 +28,47 @@ report.plugin.goals.column.isMavenReport=Report?
 report.plugin.goals.column.description=Description
 report.plugin.goal.nodescription=No description.
 report.plugin.goal.deprecated=Deprecated.
+report.plugin.goal.notice.prefix=Note
+report.plugin.goal.notice.isMavenReport=This goal should be used as a Maven report.
+report.plugin.goal.fullname=Full name
+report.plugin.goal.description=Description
+
+report.plugin.goal.attributes=Attributes
+report.plugin.goal.projectRequired=Requires a Maven project to be executed
+report.plugin.goal.aggregator=Executes as an aggregator goal
+report.plugin.goal.directInvocationOnly=Executes by direct invocation only
+report.plugin.goal.reportingMojo=Executes only as a reportSet (reporting goal)
+report.plugin.goal.dependencyResolutionRequired=Requires dependency resolution of artifacts in scope
+report.plugin.goal.dependencyCollectionRequired=Requires dependency collection of artifacts in scope
+report.plugin.goal.since=Since version
+report.plugin.goal.phase=Binds by default to the {lifecycle phase, https://maven.apache.org/ref/current/maven-core/lifecycles.html}
+report.plugin.goal.threadSafe=The goal is thread-safe and supports parallel builds
+report.plugin.goal.notThreadSafe=The goal is not marked as thread-safe and thus does not support parallel builds
+report.plugin.goal.executePhase=Invokes the execution of the following {lifecycle phase, https://maven.apache.org/ref/current/maven-core/lifecycles.html} prior to executing itself
+report.plugin.goal.executeGoal=Invokes the execution of the following plugin's goal prior to executing itself
+report.plugin.goal.executeLifecycle=Executes in its own lifecycle
+report.plugin.goal.onlineRequired=Requires that Maven runs in online mode
+report.plugin.goal.notInheritedByDefault=Is NOT inherited by default in multi-project builds
+
+report.plugin.goal.parameters=Parameters
+report.plugin.goal.noParameter=(no parameters)
+report.plugin.goal.requiredParameters=Required Parameters
+report.plugin.goal.optionalParameters=Optional Parameters
+
+report.plugin.goal.parameter.name.header=Name
+report.plugin.goal.parameter.description.header=Description
+report.plugin.goal.parameter.type.header=Type
+report.plugin.goal.parameter.since.header=Since
+report.plugin.goal.parameter.name=<{0}>
+report.plugin.goal.parameter.defaultValue=Default
+report.plugin.goal.parameter.property=User Property
+report.plugin.goal.parameter.alias=Alias
+report.plugin.goal.parameter.deprecated=<strong>Deprecated.</strong> {0}
+report.plugin.goal.parameter.type=Type
+report.plugin.goal.parameter.required=Required
+report.plugin.goal.parameter.expression=Expression
+report.plugin.goal.parameter.details=Parameter Details
+report.plugin.goal.parameter.since=Since
 
 report.plugin.systemrequirements=System Requirements
 report.plugin.systemrequirements.intro=The following specifies the minimum requirements to run this Maven plugin:
@@ -49,3 +90,5 @@ report.plugin.configuration.end=For more information, see {"Guide to Configuring
 
 report.plugin.isReport=Yes
 report.plugin.isNotReport=No
+report.plugin.yes=Yes
+report.plugin.no=No
