+++ maven-plugin-tools-api/src/test/java/org/apache/maven/tools/plugin/util/stubs/MavenReportStub.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.maven.tools.plugin.generator.stubs;
+package org.apache.maven.tools.plugin.util.stubs;
 
 import java.util.Locale;
